settings: nvs: Fix first time write issue with cache
When the setting nvs cache is disabled and `settings_nvs_save` is
called, the function runs through all name ids until either finds a
name id with the desired name or reaches the last stored name id.

With the settings nvs cache enabled, `settings_nvs_save` runs through
the cached name ids first. If the cached name matches with the desired
one, it immediately reads the stored value.

However, if the name id was not found in the cache, `settings_nvs_save`
will run through all stored name ids again. This means that until the
number of stored entries in the settings is less than the cache size,
and a new settings entry is to be stored, `settings_nvs_save` will first
run through cache, then will run through all stored name ids and only
then will pick the next free name id and finally store the new settings
entry.

This makes sence when the cache is full, but when it is known that all
entries are loaded from NVS, checking the cache is sufficient.

In the Bluetooth mesh there is a Replay Protection List which keeps
sequence numbers of all nodes it received messages from. The RPL is
stored persistently. The key of the settings entry is the source address
of the node. The common use case is when RPL is quite big (for
example, 255 entries).

With the current settings nvs cache implementation, every time the node
stores a new entry (first message from a particular source address),
`settings_nvs_save` will always check the cache first and also read all
stored entries until figures out that this is a new entry. This behavior
results in much worse (up to 10 - 20 times) performance in comparison
with when the corresponding entry was already stored.

This commit improves this behavior by bypassing the reading of NVS if
the following conditions are met:
1. `settings_nvs_load` was called,
2. the cache was not overflowed (bigger than number of stored entries).
As long as these 2 conditiones are met, it is safe to skip reading,
pick the next free name id and write the value immediately.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>